requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0

langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
openai>=1.0.0
tiktoken>=0.7.0
faiss-cpu>=1.7.0

pandas>=2.2.2
numpy>=1.26.4
reportlab>=4.0.0

pytest>=7.0.0
pytest-asyncio>=0.21.0